requirements.txt
# Core OCR and PDF processing
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80

# Vector store and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.3.1
numpy==1.24.3

# LLM integration
ollama==0.1.6

# Text processing
tqdm==4.66.1

# Web framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Document export
reportlab==4.0.9
python-docx==1.1.0


